multiple animations working and interpolation between same animation

+++ src/ui.cpp
@@ -109,7 +109,7 @@ unsigned int UIManager::setUpFont()
             texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x};
+            static_cast<unsigned int>(face->glyph->advance.x)};
         characters.insert(std::pair<char, Character>(c, character));
     }
